+++ packages/playwright-vscode-ext/src/index.ts
@@ -30,7 +30,8 @@ export {
   ensureSecondarySideBarHidden,
   waitForExtensionsActivated,
   closeWorkspaceToEmptyWindow,
-  prepareNoFolderOpenForPaletteTests
+  prepareNoFolderOpenForPaletteTests,
+  disableMonacoAutoClosing
 } from './utils/workflows';
 
 export { activeQuickInputWidget, activeQuickInputTextField } from './utils/quickInput';
@@ -84,6 +85,31 @@ export {
   verifyCommandExists
 } from './pages/commands';
 export type { OpenCommandPaletteOptions } from './pages/commands';
+
+export {
+  focusOnFilesExplorer,
+  newUntitledTextFile,
+  saveFile,
+  clearAllNotifications,
+  closeAllEditors,
+  showExplorer,
+  reloadWindow,
+  goToFile,
+  goToLineColumn,
+  selectAll,
+  paste,
+  focusOnProblemsView,
+  goToDefinition,
+  focusActiveEditorGroup,
+  find,
+  hidePanel,
+  clearOutput,
+  closeEditor,
+  insertSnippet,
+  showRunningExtensions,
+  hideSecondarySideBar,
+  closeWorkspace
+} from './pages/nativeCommands';
 
 export { executeEditorContextMenuCommand, executeExplorerContextMenuCommand } from './pages/contextMenu';